WIP Rope::remove() implementation.
It is functioning properly in that the Rope is valid and the
correct bits of text are removed.  However, it doesn't do the
required node merging to the keep the child count of nodes
above the minimum, which essentially means that the tree
just gets badly fragmented.  Implementing that is next on
the list.