Made RopeBuilder ~2x as fast.
The biggest bottleneck was the NodeText::from_str() method, which
was implemented very inefficiently.  Making that efficient was
pretty low-hanging fruit.

The new bottleneck is the Node::text_info() method for leaf nodes,
which is also the bottleneck for most of Rope's operations. Making
that faster will make a lot of things faster.