Stop-gap to make the grapheme seam corner-case work.
With the current code, we ocassionally need to spill into heap
memory when fixing CRLF grapheme seams.  So I'm commenting out
the asserts that ensure that doesn't happen.  I've left a TODO
to uncomment them once the corner-case has been addressed.

Addressing the corner case, however, is turning out to be
non-trivial, so it may have to wait for a while.  This at least
makes everything work properly.  And the heap spill is extremely
rare (to the extent that I had to pains takingly manually build
a tree to trigger it--I couldn't get it to trigger through normal
rope manipulation in practice).