Line counting improvements.
The word-at-a-time line break counting can now count all of the
unicode line breaks, and never has to fall back to the byte-at-a-time
counting because of line break types.

At the moment this doesn't produce much speedup over the previous
code, although it should lessen or eliminate the negative
performance impact with text in the higher unicode values near
the higher value line break code points.