Faster implementation of Rope::insert().
The old implementation based on 'edit_char_range()' was just a
mess and hard to follow.  This new one is not only easier to
follow, but also faster, and doesn't require outside CRLF fixing.

I expect I'll be doing the same with Rope::remove() soon, for
simpler implementation and hopefully performance improvements.