Cleaned up the behavior of the new `bytes_at()` etc. methods.
Things are better specified and more consistent now with the rest
of the APIs.

Also, the `Chunks` iterator has been changed to yield no chunks
if the `Rope` or `RopeSlice` it was created from is empty.  This
is necessary to make the semantics of the `chunks_at_*()` methods
consistent.  It was ambiguous in the documented API what the
behavior was in the this case, so I've also updated the
documentation to be explicit about it.  I don't believe this
change warrants a major version bump, as the previous behavior
wasn't clear, and this is very unlikely to break real code.