feat: cache progression in the `Lines` iterator, making around 30% faster
Constructing the lines for the `Rope::lines()` iterator used to compute lots of information twice:
when moving forward, the end of a line is always the start of the next one (and conversely when
moving backwards). Keeping this information in cache means the computation is only done once.

Further improvements can be done on the computation itself probably.