Fix Hash impl on Rope and RopeSlice.
The previous impl misunderstood the Hasher API, and assumed that
as long as the same data was passed in the same order, the hash
output would be the same.  But in fact, the data must be split
among the `write()` calls exactly the same as well.  This new impl
fixes that by always passing the data in exact fixed-sized blocks,
regardless of the underlying chunks boundaries of the rope.

Co-authored-by: Pascal Kuthe <pascal.kuthe@semimod.de>