Replace gap buffer with a flat buffer.
After actually exploring this, it's become clear that gap buffers at
the leaf nodes forces some tradeoffs that I just don't want to make,
as well as increasing the complexity of the code base.  This does mean
that best-case editing performance won't be as fast, but it also means
that the non-editing APIs will be simpler and likely a bit faster.