Add `ChunkCursor` type.
Making this its own separate type that doesn't have to adhere to the
iterator APIs allows us to present a very simple, intuitive model for
working with chunks.  The `Chunks` iterator will still exist, but will
be implemented in terms of this in a future commit, along with the
`Bytes` and `Chars` iterators.