Fix `BackingArray` undefined behavior
`BackingArray` unsafe-impls `smallvec::Array`, which represents a
promise to the smallvec crate that a raw pointer to BackingArray can be
casted into a raw pointer to a u8. However, without this commit, this
is undefined behavior, as the layout of Rust structs is unspecified
unless a `repr` annotation is used to ensure a specific layout.

This commit fixed this problem by marking `BackingArray` as
`repr(transparent)` to ensure it has the same layout as
`[u8; MAX_BYTES]`.

This also fixes the `field `0` is never read` warning that previously
occurred when running `cargo check`.