Ropey 2.0: Add support for constructing `RopeSlice`s from `&str`s. (#104)
In Ropey 1.x you could do `"Some string
slice".into()` to build a `RopeSlice` from a plain
string slice, but this wasn't implemented in Ropey
2.0.

The rationale for omitting this feature from Ropey
2.0 was:

1. Rope slices aren't string slices.
2. The initial reason for implementing this in
   Ropey 1.x was actually just to allow client
   code to do certain esoteric optimizations (as
   noted in the documentation), and was never
   intended as a generally useful feature.
3. `RopeSlice`s constructed this way are a
   significant performance footgun if the `&str`s
   they are constructed from are long, because the
   performance of almost all operations on
   `&str`-derived `RopeSlice`s is actually O(N)
   rather than the documented O(log N).

Nevertheless, in practice it's turned out to be
useful to do this in a variety of circumstances,
and some code bases have come to rely on this
feature quite deeply in their code.  Moreover,
it's the only way to construct `'static`
`RopeSlice`s, which is needed for them to be a
true analog to `&str` for `Rope`s.

Therefore, this PR adds this feature to Ropey 2.0,
but with a warning in the documentation about its
performance implications with long `&str`.