Add a `CharIndices` cursor-like Iterator
`CharIndices` can be useful in some cases where you want the byte index
of a char. It can be created on top of `Chars` outside of this crate but
within this crate we can reuse `Chars`' current byte index information
so that `CharIndices` does not need to track this info itself.

`CharIndices` mirrors `Bytes` and `Chars` in being cursor-like: it can
be reversed and you can use `next` and `prev` to go forward and backward
(depending on whether the cursor is reversed).